pyproject.toml
[tool.poetry]
name = "autotest-project-stepik"
version = "0.1.0"
description = "Autotests for Stepik website"
authors = ["Julia Ivanova <murova47@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.11"
pytest = "7.2.0"
python-dotenv = "^0.21.0"
selene = "2.0.0-beta.14"
allure-pytest = "^2.11.0"
requests = "^2.28.1"
pydantic = "^1.10.2"
pytest-xdist = "^2.5.0"
Appium-Python-Client = "^2.7.0"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
curlify = "^2.2.1"
helpers = "^0.2.0"
pytest-dependency = "^0.5.1"



[tool.pytest.ini_options]
addopts="-vv --alluredir=allure-results --clean-alluredir"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
